build boot.scr in CI
Copy the boot.cmd and gen_boot_scr.sh from backup to the top-level, and run them as part of the CI build

This ends up being a little complex because the base github vm doesn't have the tools to compile boot.cmd, and the docker image can't handle the loopfs for genimage. So we do it in two steps and copy the boot.scr around.